# build/build/swagger.json
{
    "components": {
        "examples": {},
        "headers": {},
        "parameters": {},
        "requestBodies": {},
        "responses": {},
        "schemas": {
            "CreateCursoDTO": {
                "properties": {
                    "codigo": {
                        "type": "string"
                    },
                    "imagen": {
                        "type": "string",
                        "nullable": true
                    },
                    "titulo": {
                        "type": "string"
                    }
                },
                "required": [
                    "codigo",
                    "imagen",
                    "titulo"
                ],
                "type": "object"
            },
            "UpdateCursoDTO": {
                "properties": {
                    "estado": {
                        "type": "boolean"
                    },
                    "imagen": {
                        "type": "string",
                        "nullable": true
                    },
                    "titulo": {
                        "type": "string"
                    }
                },
                "required": [
                    "estado",
                    "imagen",
                    "titulo"
                ],
                "type": "object"
            }
        },
        "securitySchemes": {}
    },
    "info": {
        "title": "backend",
        "version": "1.0.0",
        "license": {
            "name": "MIT"
        },
        "contact": {
            "name": "Loxi"
        }
    },
    "openapi": "3.0.0",
    "paths": {
        "/Curso/curso-lista": {
            "get": {
                "operationId": "GetCursoLista",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "codigo": {
                                                "type": "string"
                                            },
                                            "imagen": {
                                                "type": "string"
                                            },
                                            "titulo": {
                                                "type": "string"
                                            },
                                            "id": {
                                                "type": "number",
                                                "format": "double"
                                            }
                                        },
                                        "required": [
                                            "codigo",
                                            "imagen",
                                            "titulo",
                                            "id"
                                        ],
                                        "type": "object"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Curso"
                ],
                "security": [],
                "parameters": []
            }
        },
        "/Curso/curso/{cursoCodigo}": {
            "get": {
                "operationId": "GetCursoPorCodigo",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "codigo": {
                                            "type": "string"
                                        },
                                        "imagen": {
                                            "type": "string"
                                        },
                                        "titulo": {
                                            "type": "string"
                                        },
                                        "id": {
                                            "type": "number",
                                            "format": "double"
                                        }
                                    },
                                    "required": [
                                        "codigo",
                                        "imagen",
                                        "titulo",
                                        "id"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Curso"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "cursoCodigo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/Curso/crear-curso": {
            "post": {
                "operationId": "CrearCurso",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "codigo": {
                                            "type": "string"
                                        },
                                        "imagen": {
                                            "type": "string"
                                        },
                                        "titulo": {
                                            "type": "string"
                                        },
                                        "id": {
                                            "type": "number",
                                            "format": "double"
                                        }
                                    },
                                    "required": [
                                        "codigo",
                                        "imagen",
                                        "titulo",
                                        "id"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Curso"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCursoDTO"
                            }
                        }
                    }
                }
            }
        },
        "/Curso/actualizar-curso/{cursoCodigo}": {
            "put": {
                "operationId": "ActualizarCurso",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "estado": {
                                            "type": "boolean"
                                        },
                                        "codigo": {
                                            "type": "string"
                                        },
                                        "imagen": {
                                            "type": "string"
                                        },
                                        "titulo": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "estado",
                                        "codigo",
                                        "imagen",
                                        "titulo"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Curso"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "cursoCodigo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCursoDTO"
                            }
                        }
                    }
                }
            }
        },
        "/departamento/departamento-lista": {
            "get": {
                "operationId": "GetDepartamentoLista",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "departamento": {
                                                "type": "string"
                                            },
                                            "id": {
                                                "type": "number",
                                                "format": "double"
                                            }
                                        },
                                        "required": [
                                            "departamento",
                                            "id"
                                        ],
                                        "type": "object"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Curso"
                ],
                "security": [],
                "parameters": []
            }
        },
        "/departamento/departamento/{codigoId}": {
            "get": {
                "operationId": "GetDepartamentoPorId",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "departamento": {
                                            "type": "string"
                                        },
                                        "id": {
                                            "type": "number",
                                            "format": "double"
                                        }
                                    },
                                    "required": [
                                        "departamento",
                                        "id"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Curso"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "codigoId",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    }
                ]
            }
        },
        "/user/get-user": {
            "get": {
                "operationId": "GetUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [],
                "parameters": []
            }
        },
        "/user/add-user": {
            "post": {
                "operationId": "AddUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "edad": {
                                            "type": "number",
                                            "format": "double"
                                        },
                                        "nombre": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "edad",
                                        "nombre"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "edad": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "nombre": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "edad",
                                    "nombre"
                                ],
                                "type": "object"
                            }
                        }
                    }
                }
            }
        },
        "/user/show-user": {
            "get": {
                "operationId": "ShowUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "lenguaje": {
                                                "type": "string"
                                            },
                                            "edad": {
                                                "type": "number",
                                                "format": "double"
                                            },
                                            "sexo": {
                                                "type": "string"
                                            },
                                            "apellidos": {
                                                "type": "string"
                                            },
                                            "nombre": {
                                                "type": "string"
                                            }
                                        },
                                        "required": [
                                            "lenguaje",
                                            "edad",
                                            "sexo",
                                            "apellidos",
                                            "nombre"
                                        ],
                                        "type": "object"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "security": [],
                "parameters": []
            }
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}